requirements.txt
grpcio-tools
flask ~= 1.1
markupsafe ~= 2.0.1